Module core_extensions::slices

Slice extension traits, and related items.

Structs

KeySlice

A pair of (slice, key) returned by the RSplitSliceWhile/ SplitSliceWhile iterators.

RSplitSliceWhile

Iterator over slices, in which all the elements in each slice were mapped to the same key by a closure, iterating from the end.

SliceBias

What directions SliceExt::{slice_lossy, slice_lossy_mut} are biased towards.

SplitSliceWhile

Iterator over slices, in which all the elements in each slice were mapped to the same key by a closure.

Enums

BiasDirection

The direction a range bound is moved towards to make the bound a valid index.

Traits

SliceExt

Extension trait for [T] and str.

ValSliceExt

Extension trait for [T].