Struct core_extensions::type_asserts::AssertEq

pub struct AssertEq<L, R: ?Sized>
where
    L: TypeIdentity<Type = R> + ?Sized,
{ /* private fields */ }

Asserts that its 2 type parameters are the same type.

This assertion is done on the type level, so let _: AssertEq<T, U>; requires that T must be the same type as U.

Example

use core_extensions::type_asserts::AssertEq;

trait ElemTy {
    type Elem;
}

impl<A> ElemTy for (A,) {
    type Elem = A;
}

let _: AssertEq<u32, <(u32,) as ElemTy>::Elem>;

let _ = AssertEq::new(&0u32, &0u32);

Non-compiling

use core_extensions::type_asserts::AssertEq;

let _: AssertEq<(), u32>;
let _ = AssertEq::new(&(), &0u32);

use core_extensions::type_asserts::AssertEq;

let _: AssertEq<u32, ()>;
let _ = AssertEq::new(&0u32, &());

Implementations

impl<A> AssertEq<A, A>

pub fn new(_: A, _: A) -> Self

Constructs an AssertEq.

pub const NEW: Self = _

Constructs an AssertEq.