core_extensions::type_asserts

Struct AssertEq3

pub struct AssertEq3<A, B: ?Sized, C: ?Sized>
where
    A: TypeIdentity<Type = B> + TypeIdentity<Type = C> + ?Sized,
{ /* private fields */ }

Asserts that its 3 type parameters are the same type.

This assertion is done on the type level, so let _: AssertEq3<A, B, C>; requires that A, B, and C must be the same type.

Example

use core_extensions::type_asserts::AssertEq3;

trait TypeParams {
    type First;
    type Second;
}

impl<A, B> TypeParams for (A, B) {
    type First = A;
    type Second = B;
}

type First<T> = <T as TypeParams>::First;

type Second<T> = <T as TypeParams>::Second;

let _: AssertEq3<u32, First<(u32, ())>, Second<((), u32)>>;

let _ = AssertEq3::new(&0u32, &0u32, &0u32);

Non-compiling

let _: AssertEq3<(), u32, u32>;
let _ = AssertEq3::new(&(), &0u32, &0u32);

let _: AssertEq3<u32, (), u32>;
let _ = AssertEq3::new(&0u32, &(), &0u32);

let _: AssertEq3<u32, u32, ()>;
let _ = AssertEq3::new(&0u32, &0u32, &());

Implementations

impl<A> AssertEq3<A, A, A>

pub const NEW: Self

Constructs an AssertEq3.

pub fn new(_: A, _: A, _: A) -> Self

Constructs an AssertEq3.