Module core_extensions::strings

Extension trait for string types.

Structs

CharIndicesFrom

Like CharIndices, which starts from an offset.

KeyStr

A pair of (string slice, key) returned by the RSplitWhile/ SplitWhile iterators.

LeftPadder

Add padding to a string in its Display impl.

RSplitWhile

Iterator over string slices, in which all the chars in each string were mapped to the same key by a closure, iterating from the end.

SplitWhile

Iterator over string slices, in which all the chars in each string were mapped to the same key by a closure.

Traits

StringExt

Extension trait for strings (any type that borrows as str).